requirements.txt
# Core CLI dependencies
click>=8.1.7
colorama>=0.4.6
python-dotenv>=1.0.0
tabulate>=0.9.0

# HTTP flask
flask-socketio
psutil
loguru
requests
python-dotenv
beautifulsoup4
cryptography
boto3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
llama-index
llama-index-core
llama-index-llms-openai
pyautogen
scrapegraphai
playwright
ragas
datasets>=4.12.2
pysocks
onionsearch
firecrawl-py
scrapling
lxml>=4.9.3

# OSINT libraries
sherlock-project>=0.14.3
holehe>=1.61
socialscan>=1.4.2

# Web interface
Flask>=3.0.0
Flask-SocketIO>=5.3.5
Flask-CORS>=4.0.0
eventlet>=0.33.3

# Telegram bot
python-telegram-bot>=20.7

# Intelligence layer
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.26.2
python-magic>=0.4.27
mmh3>=3.0.0

# Data handling
openpyxl>=3.1.2
Jinja2>=3.1.2

# Async support
aiohttp>=3.9.1
asyncio>=3.4.3

# Logging and monitoring
loguru>=0.7.2
psutil>=5.9.0

# Rate limiting
ratelimit>=2.2.1

# Database (optional for future)
SQLAlchemy>=2.0.23

# Graph Analysis
networkx>=3.1
python-louvain>=0.16
